drivers: aw8624_haptic: Fix -Wpointer-to-int-cast
../drivers/input/misc/aw8624_haptic/aw8624.c:2271:8: warning: cast to smaller integer type 'int' from 'u8 *' (aka 'unsigned char *') [-Wpointer-to-int-cast]
                     (int)effect->pattern);
                     ^~~~~~~~~~~~~~~~~~~~

+++ drivers/input/misc/aw8624_haptic/aw8624.c
@@ -2637,7 +2637,7 @@ static int aw8624_parse_dt(struct device *dev, struct aw8624 *aw8624,
 		printk
 		    ("%s: %d  effect->pattern_length=%d  effect->pattern=%d \n",
 		     __func__, __LINE__, effect->pattern_length,
-		     (int)effect->pattern);
+		     (uintptr_t)effect->pattern);
 
 		effect->play_rate_us = config->play_rate_us;
 		rc = of_property_read_u32(child_node, "qcom,wf-play-rate-us",
